+++ gen/plugins/plugins_test.go
@@ -7,7 +7,8 @@ var (
 	_ gen.DBInfoPlugin[any, any, any] = &enumsPlugin[any, any, any]{}
 	_ gen.StatePlugin[any]            = modelsPlugin[any]{}
 	_ gen.StatePlugin[any]            = factoryPlugin[any]{}
-	_ gen.StatePlugin[any]            = queriesOutputPlugin[any]{}
+	_ gen.StatePlugin[any]            = &queriesOutputPlugin[any, any, any]{}
+	_ gen.DBInfoPlugin[any, any, any] = &queriesOutputPlugin[any, any, any]{}
 	_ gen.StatePlugin[any]            = dbErrorsPlugin[any]{}
 	_ gen.StatePlugin[any]            = joinsPlugin[any]{}
 	_ gen.StatePlugin[any]            = loadersPlugin[any]{}