+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `enums`: Generates code for enums in a separate package, if there are any present.
   - `models`: Generates code for models. Depends on `enums`.
   - `factory`: Generates code for factories. Depends on `models`.
-  - `queries`: Generates code for queries. Depends on `enums`.
+  - `queries`: Generates code for queries.
   - `dberrors`: Generates code for unique constraint errors. Depends on `models`.
   - `joins`: Adds templates to the `models` package to generate code for joins e.g `models.SelectJoin.Table.LeftJoin.Rel`.
   - `loaders`: Adds templates to the `models` package to generate code for loaders e.g `models.SelectThenLoad.Table.Rel()`.